chore: Break out `Button.tsx` file
- Removed the old `Ripple` component as we plan on phasing out use of it.
