+++ mobile/src/navigation/components/MiniPlayer.tsx
@@ -12,7 +12,7 @@ import { usePreferenceStore } from "~/stores/Preference/store";
 import { OnRTL } from "~/lib/react";
 import { cn } from "~/lib/style";
 import { Marquee } from "~/components/Containment/Marquee";
-import { IconButton } from "~/components/Form/Button";
+import { IconButton } from "~/components/Form/Button/Icon";
 import { Swipeable } from "~/components/Swipeable";
 import { StyledText } from "~/components/Typography/StyledText";
 import {
