+++ mobile/src/navigation/components/TopAppBar.tsx
@@ -9,7 +9,7 @@ import { ArrowBack } from "~/resources/icons/ArrowBack";
 
 import { OnRTL } from "~/lib/react";
 import { SafeContainer } from "~/components/Containment/SafeContainer";
-import { IconButton } from "~/components/Form/Button";
+import { IconButton } from "~/components/Form/Button/Icon";
 import { StyledText } from "~/components/Typography/StyledText";
 
 /**
