+++ mobile/src/navigation/screens/HomeView.tsx
@@ -11,7 +11,8 @@ import { StandardScrollLayout } from "../layouts/StandardScroll";
 import { cn } from "~/lib/style";
 import { abbreviateNum } from "~/utils/number";
 import { LegendList } from "~/components/Defaults";
-import { Button, IconButton } from "~/components/Form/Button";
+import { Button } from "~/components/Form/Button";
+import { IconButton } from "~/components/Form/Button/Icon";
 import { AccentText } from "~/components/Typography/AccentText";
 import { TEm, TStyledText } from "~/components/Typography/StyledText";
 import {
