+++ mobile/src/navigation/screens/now-playing/UpcomingView.tsx
@@ -21,8 +21,8 @@ import { debounceWithAccumulation } from "~/utils/debounce";
 import { moveArray } from "~/utils/object";
 import { wait } from "~/utils/promise";
 import { FlashDragList } from "~/components/Defaults";
-import type { PressProps } from "~/components/Form/Button";
-import { IconButton } from "~/components/Form/Button";
+import { IconButton } from "~/components/Form/Button/Icon";
+import type { PressProps } from "~/components/Form/Button/types";
 import { Swipeable } from "~/components/Swipeable";
 import { PlayingIndicator } from "~/modules/media/components/AnimatedBars";
 import { SearchResult } from "~/modules/search/components/SearchResult";
