+++ mobile/src/navigation/screens/playlists/ModifyViewBase/index.tsx
@@ -30,7 +30,8 @@ import { cn } from "~/lib/style";
 import { ToastOptions } from "~/lib/toast";
 import { wait } from "~/utils/promise";
 import { FlashDragList } from "~/components/Defaults";
-import { Button, IconButton } from "~/components/Form/Button";
+import { Button } from "~/components/Form/Button";
+import { IconButton } from "~/components/Form/Button/Icon";
 import { TextInput } from "~/components/Form/Input";
 import { ModalTemplate } from "~/components/Modal";
 import { useSheetRef } from "~/components/Sheet/useSheetRef";
